calculs ok sans gestion d'exception

+++ fctnsOperandes.js
@@ -1,23 +1,23 @@
 function sum(a,b) {
-    a = parseInt(a);
-    b = parseInt(b);
+    a = parseFloat(a);
+    b = parseFloat(b);
     return (a+b);
 }
 
 function minus(a,b) {
-    a = parseInt(a);
-    b = parseInt(b);
+    a = parseFloat(a);
+    b = parseFloat(b);
     return (a-b);
 }
 
 function multiply(a,b) {
-    a = parseInt(a);
-    b = parseInt(b);
+    a = parseFloat(a);
+    b = parseFloat(b);
     return (a*b);
 }
 
 function divide(a,b) {
-    a = parseInt(a);
-    b = parseInt(b);
+    a = parseFloat(a);
+    b = parseFloat(b);
     return (a/b);
 }
